locate: change updatedb invocation for mlocate
The locate service does not properly handle flags during update of the
relevant database when configured to use the mlocate package.

The manual entry for updatedb associated with mlocate does not say
that it supports environment variables in place of command line flags,
whereas the `findutils` package's updatedb does so.

Fixes #29279